refactor: lift state into app
